Update StudyTracker.java
fixed what I think was a casting error

+++ src/main/java/edu/touro/las/mcon364/test/StudyTracker.java
@@ -60,7 +60,7 @@ public class StudyTracker {
         }
         var scores = scoresByLearner.get(name);
         scores.add(score);
-        undoStack.push(() -> scoresByLearner.get(name).remove(score));
+        undoStack.push(() -> scoresByLearner.get(name).remove((Integer) score));
         return true;
     }
